requirements.txt
certifi==2021.5.30
cycler==0.10.0
future==0.18.2
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.17.0
# nvidia-dali-cuda100==1.3.0
onnx==1.6.0
onnxruntime==1.3.0
opencv-python==4.2.0.34
parse==1.19.0
Pillow==8.3.2
protobuf==3.18.0
pyparsing==2.4.7
python-dateutil==2.8.2
six==1.16.0
torch==1.6.0
torchvision==0.7.0
typing-extensions==3.10.0.2
